Refactor Camel K routing and use BigInteger for precision

+++ fee-service-eu/src/main/kotlin/com/example/strata/fee/controller/FeeController.kt
@@ -6,9 +6,9 @@ import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
-import java.math.BigDecimal
+import java.math.BigInteger
 
-data class FeeRequest(val amount: BigDecimal)
+data class FeeRequest(val amount: BigInteger)
 
 @RestController
 @RequestMapping("/api/fees")
